Went on some huge Scala detours
In order to map a function over a list of tuples, I needed to use a case
statement in some strange way to properly pass in the parameters.

I'm doing some serious hoop jumping to stay within the bounds of
immutable vectors, so I'm not sure if I'm actually gaining anything with
this approach, or if I'm torturing these immutable structures so much
that I should be using the mutable versions instead.